End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,13 +17,13 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
 ngOnInit(): void {
   this.loadMessages();
 }
 
-  loadMessages(): void {
+  loadMessages() {
     this.loading = true;
     this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response =>{
       this.messages = response.result;
@@ -33,8 +34,14 @@ ngOnInit(): void {
   }
 
   deleteMessage(id: number){
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result =>{
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {                                      // moved to here  // allows for confirmation to be displayed before we delete a message
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    })
+  }
+    // this.messageService.deleteMessage(id).subscribe(() => {
+    //   this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
     })
   }
